app/requirements.txt
Adafruit-Blinka==8.66.2
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-ltr390==1.1.21
adafruit-circuitpython-register==1.11.0
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.2
Adafruit-PlatformDetect==3.84.1
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
bme280==0.7
board==1.0
colorzero==2.0
gpiozero==2.0.1
pyftdi==0.57.1
pyserial==3.5
pyusb==1.3.1
PyYAML==6.0.3
setuptools==80.9.0
smbus2==0.5.0
sps30==0.1.10
typing_extensions==4.15.0
